adjust makefile to generate classic payment in single class
